pyproject.toml
[tool.poetry]
name = "ovdsearch"
version = "0.1.0"
description = "Bot for locating people in OVD"
authors = ["Petr Ogarok", "Arseniy Sotskiy", "Igor Ovsyannikov"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
sqlalchemy = "^1.3.11"
click = "^7.0"
python-telegram-bot = "^12.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
